lab4/zad3: complete minimal spanning tree finding

+++ lab4/zad3/Lab04.cs
@@ -8,7 +8,7 @@ class Example
     public static void Main()
     {
         bool b;
-        int mst;
+        double mst;
         Graph g, t;
         Edge[] ep;
         var rgg = new RandomGraphGenerator();
